# Speed up diff computation for lists with shared pre- or suffix