This excludes the shared prefix and suffix from the matrix, reducing computation time and memory requirements for many common cases.

Before:
```
running 7 tests
test sup::sup::slice::10 equal items       ... bench:         810 ns/iter (+/- 52)
test sup::sup::slice::10 non-equal items   ... bench:         909 ns/iter (+/- 61)
test sup::sup::slice::100 equal items      ... bench:      23,953 ns/iter (+/- 1,833)
test sup::sup::slice::100 non-equal items  ... bench:      36,420 ns/iter (+/- 4,247)
test sup::sup::slice::1000 equal items     ... bench:   3,215,529 ns/iter (+/- 316,083)
test sup::sup::slice::1000 non-equal items ... bench:   4,519,959 ns/iter (+/- 382,556)
test sup::sup::slice::empty                ... bench:          48 ns/iter (+/- 12)
```

After:
```
running 7 tests
test sup::sup::slice::10 equal items       ... bench:         114 ns/iter (+/- 11)
test sup::sup::slice::10 non-equal items   ... bench:         948 ns/iter (+/- 75)
test sup::sup::slice::100 equal items      ... bench:         435 ns/iter (+/- 31)
test sup::sup::slice::100 non-equal items  ... bench:      41,591 ns/iter (+/- 2,333)
test sup::sup::slice::1000 equal items     ... bench:       3,686 ns/iter (+/- 468)
test sup::sup::slice::1000 non-equal items ... bench:   4,947,898 ns/iter (+/- 735,268)
test sup::sup::slice::empty                ... bench:          59 ns/iter (+/- 11)
```